Update habits listing when app becomes active
