+++ Habit-Calendar/Habit-Calendar/Controllers/HabitsTableViewController/HabitsTableViewController.swift
@@ -11,7 +11,7 @@ import CoreData
 import UserNotifications
 
 /// Controller in charge of displaying the list of tracked habits.
-class HabitsTableViewController: UITableViewController, NSFetchedResultsControllerDelegate {
+class HabitsTableViewController: UITableViewController {
 
     // MARK: Types
 
@@ -24,16 +24,16 @@ class HabitsTableViewController: UITableViewController, NSFetchedResultsControll
     // MARK: Properties
 
     /// The identifier for the habit creation controller's segue.
-    private let newHabitSegueIdentifier = "Create a new habit"
+    let newHabitSegueIdentifier = "Create a new habit"
 
     /// The identifier for the habit details controller's segue.
-    private let detailsSegueIdentifier = "Show habit details"
+    let detailsSegueIdentifier = "Show habit details"
 
     /// The in progress habit cell's reuse identifier.
-    private let inProgressHabitCellIdentifier = "In progress habit table view cell"
+    let inProgressHabitCellIdentifier = "In progress habit table view cell"
 
     /// The completed habit cell's reuse identifier.
-    private let completedHabitCellIdentifier = "Completed habit table view cell"
+    let completedHabitCellIdentifier = "Completed habit table view cell"
 
     /// The used persistence container.
     var container: NSPersistentContainer!
@@ -48,7 +48,7 @@ class HabitsTableViewController: UITableViewController, NSFetchedResultsControll
     @IBOutlet weak var habitsSegmentedControl: UISegmentedControl!
 
     /// The fetched results controller used to get the habits that are in progress and display them with the tableView.
-    private lazy var progressfetchedResultsController: NSFetchedResultsController<HabitMO> = {
+    lazy var progressfetchedResultsController: NSFetchedResultsController<HabitMO> = {
         let fetchedController = habitStorage.makeFetchedResultsController(context: container.viewContext)
         fetchedController.delegate = self
 
@@ -56,7 +56,7 @@ class HabitsTableViewController: UITableViewController, NSFetchedResultsControll
     }()
 
     /// The fetched results controller used to get the habits that are completed and display them with the tableView.
-    private lazy var completedfetchedResultsController: NSFetchedResultsController<HabitMO> = {
+    lazy var completedfetchedResultsController: NSFetchedResultsController<HabitMO> = {
         let fetchedController = habitStorage.makeCompletedFetchedResultsController(context: container.viewContext)
         fetchedController.delegate = self
 
@@ -64,14 +64,14 @@ class HabitsTableViewController: UITableViewController, NSFetchedResultsControll
     }()
 
     /// The currently selected segment.
-    private var selectedSegment: Segment {
+    var selectedSegment: Segment {
         return Segment(rawValue: habitsSegmentedControl.selectedSegmentIndex)!
     }
 
     /// The fetched results controller for the selected segment (in progress or completed habits).
     /// - Note: This is the fetched results controller used by the tableView's data source, which is chosen based
     ///         on the currently selected segmented.
-    private var selectedFetchedResultsController: NSFetchedResultsController<HabitMO> {
+    var selectedFetchedResultsController: NSFetchedResultsController<HabitMO> {
         switch selectedSegment {
         case .inProgress:
             return progressfetchedResultsController
@@ -106,6 +106,13 @@ class HabitsTableViewController: UITableViewController, NSFetchedResultsControll
             self,
             selector: #selector(handleContextChanges(notification:)),
             name: Notification.Name.NSManagedObjectContextDidSave,
+            object: nil
+        )
+        // Register to notifications informing that the app has become active. Update the fetched controllers.
+        NotificationCenter.default.addObserver(
+            self,
+            selector: #selector(updateFetchedResultsController(notification:)),
+            name: Notification.Name.UIApplicationDidBecomeActive,
             object: nil
         )
 
@@ -187,10 +194,7 @@ class HabitsTableViewController: UITableViewController, NSFetchedResultsControll
     // MARK: Actions
 
     @IBAction func switchSegment(_ sender: UISegmentedControl) {
-        // TODO: Alert the user in case of errors.
-        try? selectedFetchedResultsController.performFetch()
-        displayEmptyStateIfNeeded()
-        tableView.reloadData()
+        updateList()
     }
 
     @objc private func createNewHabit() {
@@ -212,7 +216,19 @@ class HabitsTableViewController: UITableViewController, NSFetchedResultsControll
         container.viewContext.mergeChanges(fromContextDidSave: notification)
     }
 
+    @objc private func updateFetchedResultsController(notification: Notification) {
+        updateList()
+    }
+
     // MARK: Imperatives
+
+    /// Updates the controller's list of habits.
+    private func updateList() {
+        // TODO: Alert the user in case of errors.
+        try? selectedFetchedResultsController.performFetch()
+        displayEmptyStateIfNeeded()
+        tableView.reloadData()
+    }
 
     /// Instantiates a new EmptyStateView for usage.
     /// - Returns: The instantiated EmptyStateView.
@@ -229,7 +245,7 @@ class HabitsTableViewController: UITableViewController, NSFetchedResultsControll
     }
 
     /// Display the controller's empty state depending on the user's added habits.
-    private func displayEmptyStateIfNeeded() {
+    func displayEmptyStateIfNeeded() {
         // Check if the user has any habits, if he doesn't, display the empty state.
         if let count = try? container.viewContext.count(for: HabitMO.fetchRequest()), count == 0 {
             habitsSegmentedControl.isHidden = true
@@ -278,136 +294,5 @@ class HabitsTableViewController: UITableViewController, NSFetchedResultsControll
         UserDefaults.standard.setFirstLaunchPassed()
         // Present it on top of the window's root controller.
         present(presentationController, animated: true)
-    }
-
-    // MARK: DataSource Methods
-
-    override func numberOfSections(in tableView: UITableView) -> Int {
-        return selectedFetchedResultsController.sections?.count ?? 1
-    }
-
-    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        if let sections = selectedFetchedResultsController.sections, sections.count > 0 {
-            return sections[section].numberOfObjects
-        }
-
-        return 0
-    }
-     override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        var cell: UITableViewCell? = nil
-
-        // Get the current habit object.
-        let habit = selectedFetchedResultsController.object(at: indexPath)
-
-        switch selectedSegment {
-        case .inProgress:
-            cell = tableView.dequeueReusableCell(
-                withIdentifier: inProgressHabitCellIdentifier,
-                for: indexPath
-            )
-            if let cell = cell as? HabitTableViewCell {
-                // Display the habit properties:
-                // Its name.
-                cell.nameLabel?.text = habit.name
-
-                // And its progress.
-                let progress = habit.getCurrentChallenge()?.getCompletionProgress() ?? (0, 0)
-                cell.progressLabel?.text = "\(progress.0) / \(progress.1) completed days"
-                cell.progressBar.tint = habit.getColor().uiColor
-                // Change the bar's progress (past days / total).
-                cell.progressBar.progress = CGFloat(Double(progress.0) / Double(progress.1))
-            }
-        case .completed:
-            cell = tableView.dequeueReusableCell(
-                withIdentifier: completedHabitCellIdentifier,
-                for: indexPath
-            )
-            if let cell = cell as? CompletedHabitTableViewCell {
-                // Display the habit's name and color.
-                cell.nameLabel.text = habit.name
-                cell.colorView.backgroundColor = habit.getColor().uiColor
-            }
-        }
-
-        return cell!
-    }
-
-    // MARK: Delegate methods
-
-    override func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        switch selectedSegment {
-        case .inProgress:
-            return 145
-        case .completed:
-            return 100
-        }
-    }
-
-    override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        performSegue(withIdentifier: detailsSegueIdentifier, sender: self)
-    }
-}
-
-extension HabitsTableViewController {
-
-    // MARK: NSFetchedResultsControllerDelegate implementation methods
-
-    func controllerWillChangeContent(_ controller: NSFetchedResultsController<NSFetchRequestResult>) {
-        // Begin the tableView updates.
-        tableView.beginUpdates()
-    }
-
-    func controller(
-        _ controller: NSFetchedResultsController<NSFetchRequestResult>,
-        didChange anObject: Any,
-        at indexPath: IndexPath?,
-        for type: NSFetchedResultsChangeType,
-        newIndexPath: IndexPath?
-    ) {
-        // Only execute the updates if the segment being shown is handled by the passed controller.
-        // There're two fetchedResultsController instances managing different segments of the table view,
-        // One for the in progress habits and another for the completed ones. Only update the one being displayed.
-        guard shouldUpdateSegmentFrom(controller: controller) else { return }
-
-        // Add or remove rows based on the kind of changes:
-        switch type {
-        case .delete:
-            if let indexPath = indexPath {
-                tableView.deleteRows(at: [indexPath], with: .automatic)
-            }
-        case .insert:
-            if let indexPath = newIndexPath {
-                tableView.insertRows(at: [indexPath], with: .automatic)
-            }
-        case .move:
-            if let indexPath = indexPath, let newIndexPath = newIndexPath {
-                tableView.deleteRows(at: [indexPath], with: .automatic)
-                tableView.insertRows(at: [newIndexPath], with: .automatic)
-            }
-        case .update:
-            if let indexPath = indexPath {
-                tableView.reloadRows(at: [indexPath], with: .automatic)
-            }
-        }
-    }
-
-    /// Informs if the changed fetched results controller should update the current segment being displayed
-    /// by the table view.
-    /// - Parameter fetchedResultsController: The updated fetched results controller.
-    /// - Returns: True if the update should be displayed by the segment, false otherwise.
-    private func shouldUpdateSegmentFrom(controller: NSFetchedResultsController<NSFetchRequestResult>) -> Bool {
-        switch selectedSegment {
-        case .inProgress:
-            return controller == progressfetchedResultsController
-        case .completed:
-            return controller == completedfetchedResultsController
-        }
-    }
-
-    func controllerDidChangeContent(_ controller: NSFetchedResultsController<NSFetchRequestResult>) {
-        // End the tableView updates.
-        tableView.endUpdates()
-
-        displayEmptyStateIfNeeded()
     }
 }
